working

+++ src/mathLib/vec3.cpp
@@ -3,7 +3,7 @@
 using namespace mathLib;
 
 
-vec3::vec3(const float e0, const float e1, const float e2)
+vec3::vec3(const double e0, const double e1, const double e2)
 {
 	e[0] = e0;
 	e[1] = e1;
